chore/destination_elder: extract out MDataHandler
