+++ src/destination_elder.rs
@@ -15,6 +15,7 @@ use crate::{
         error::Error as ChunkStoreError, AppendOnlyChunkStore, ImmutableChunkStore,
         MutableChunkStore,
     },
+    mdata_handler::MDataHandler,
     rpc::Rpc,
     utils,
     vault::Init,
@@ -58,7 +59,7 @@ pub(crate) struct DestinationElder {
     immutable_metadata: PickleDb,
     full_adults: PickleDb,
     immutable_chunks: ImmutableChunkStore,
-    mutable_chunks: MutableChunkStore,
+    mdata_handler: MDataHandler,
     adata_handler: ADataHandler,
 }
 
@@ -80,12 +81,7 @@ impl DestinationElder {
             Rc::clone(total_used_space),
             init_mode,
         )?;
-        let mutable_chunks = MutableChunkStore::new(
-            &root_dir,
-            max_capacity,
-            Rc::clone(total_used_space),
-            init_mode,
-        )?;
+        let mdata_handler = MDataHandler::new(id.clone(), config, total_used_space, init_mode)?;
         let adata_handler = ADataHandler::new(id.clone(), config, total_used_space, init_mode)?;
         Ok(Self {
             id,
@@ -93,7 +89,7 @@ impl DestinationElder {
             immutable_metadata,
             full_adults,
             immutable_chunks,
-            mutable_chunks,
+            mdata_handler,
             adata_handler,
         })
     }
@@ -143,39 +139,45 @@ impl DestinationElder {
             //
             // ===== Mutable Data =====
             //
-            PutMData(data) => self.handle_put_mdata_req(requester, data, message_id),
-            GetMData(address) => self.handle_get_mdata_req(requester, address, message_id),
-            GetMDataValue { address, ref key } => {
-                self.handle_get_mdata_value_req(requester, address, key, message_id)
-            }
-            DeleteMData(address) => self.handle_delete_mdata_req(requester, address, message_id),
-            GetMDataShell(address) => {
-                self.handle_get_mdata_shell_req(requester, address, message_id)
-            }
-            GetMDataVersion(address) => {
-                self.handle_get_mdata_version_req(requester, address, message_id)
-            }
-            ListMDataEntries(address) => {
-                self.handle_list_mdata_entries_req(requester, address, message_id)
-            }
-            ListMDataKeys(address) => {
-                self.handle_list_mdata_keys_req(requester, address, message_id)
-            }
-            ListMDataValues(address) => {
-                self.handle_list_mdata_values_req(requester, address, message_id)
-            }
-            ListMDataPermissions(address) => {
-                self.handle_list_mdata_permissions_req(requester, address, message_id)
-            }
-            ListMDataUserPermissions { address, user } => {
-                self.handle_list_mdata_user_permissions_req(requester, address, user, message_id)
-            }
+            PutMData(data) => self
+                .mdata_handler
+                .handle_put_mdata_req(requester, data, message_id),
+            GetMData(address) => self
+                .mdata_handler
+                .handle_get_mdata_req(requester, address, message_id),
+            GetMDataValue { address, ref key } => self
+                .mdata_handler
+                .handle_get_mdata_value_req(requester, address, key, message_id),
+            DeleteMData(address) => self
+                .mdata_handler
+                .handle_delete_mdata_req(requester, address, message_id),
+            GetMDataShell(address) => self
+                .mdata_handler
+                .handle_get_mdata_shell_req(requester, address, message_id),
+            GetMDataVersion(address) => self
+                .mdata_handler
+                .handle_get_mdata_version_req(requester, address, message_id),
+            ListMDataEntries(address) => self
+                .mdata_handler
+                .handle_list_mdata_entries_req(requester, address, message_id),
+            ListMDataKeys(address) => self
+                .mdata_handler
+                .handle_list_mdata_keys_req(requester, address, message_id),
+            ListMDataValues(address) => self
+                .mdata_handler
+                .handle_list_mdata_values_req(requester, address, message_id),
+            ListMDataPermissions(address) => self
+                .mdata_handler
+                .handle_list_mdata_permissions_req(requester, address, message_id),
+            ListMDataUserPermissions { address, user } => self
+                .mdata_handler
+                .handle_list_mdata_user_permissions_req(requester, address, user, message_id),
             SetMDataUserPermissions {
                 address,
                 user,
                 ref permissions,
                 version,
-            } => self.handle_set_mdata_user_permissions_req(
+            } => self.mdata_handler.handle_set_mdata_user_permissions_req(
                 requester,
                 address,
                 user,
@@ -187,15 +189,15 @@ impl DestinationElder {
                 address,
                 user,
                 version,
-            } => self.handle_del_mdata_user_permissions_req(
+            } => self.mdata_handler.handle_del_mdata_user_permissions_req(
                 requester, address, user, version, message_id,
             ),
-            MutateSeqMDataEntries { address, actions } => {
-                self.handle_mutate_seq_mdata_entries_req(requester, address, actions, message_id)
-            }
-            MutateUnseqMDataEntries { address, actions } => {
-                self.handle_mutate_unseq_mdata_entries_req(requester, address, actions, message_id)
-            }
+            MutateSeqMDataEntries { address, actions } => self
+                .mdata_handler
+                .handle_mutate_seq_mdata_entries_req(requester, address, actions, message_id),
+            MutateUnseqMDataEntries { address, actions } => self
+                .mdata_handler
+                .handle_mutate_unseq_mdata_entries_req(requester, address, actions, message_id),
             //
             // ===== Append Only Data =====
             //
@@ -391,464 +393,6 @@ impl DestinationElder {
                 None
             }
         }
-    }
-
-    /// Get `MData` from the chunk store and check permissions.
-    /// Returns `Some(Result<..>)` if the flow should be continued, returns
-    /// `None` if there was a logic error encountered and the flow should be
-    /// terminated.
-    fn get_mdata_chunk(
-        &mut self,
-        address: &MDataAddress,
-        requester: &PublicId,
-        action: MDataAction,
-    ) -> Option<NdResult<MData>> {
-        let requester_pk = if let Some(pk) = utils::own_key(&requester) {
-            pk
-        } else {
-            error!("Logic error: requester {:?} must not be Node", requester);
-            return None;
-        };
-
-        Some(
-            self.mutable_chunks
-                .get(&address)
-                .map_err(|e| match e {
-                    ChunkStoreError::NoSuchChunk => NdError::NoSuchData,
-                    error => error.to_string().into(),
-                })
-                .and_then(move |mdata| {
-                    mdata
-                        .check_permissions(action, *requester_pk)
-                        .map(move |_| mdata)
-                }),
-        )
-    }
-
-    /// Get MData from the chunk store, update it, and overwrite the stored chunk.
-    fn mutate_mdata_chunk<F>(
-        &mut self,
-        address: &MDataAddress,
-        requester: PublicId,
-        message_id: MessageId,
-        mutation_fn: F,
-    ) -> Option<Action>
-    where
-        F: FnOnce(MData) -> NdResult<MData>,
-    {
-        let result = self
-            .mutable_chunks
-            .get(address)
-            .map_err(|e| match e {
-                ChunkStoreError::NoSuchChunk => NdError::NoSuchData,
-                error => error.to_string().into(),
-            })
-            .and_then(mutation_fn)
-            .and_then(move |mdata| {
-                self.mutable_chunks
-                    .put(&mdata)
-                    .map_err(|error| error.to_string().into())
-            });
-
-        Some(Action::RespondToClientHandlers {
-            sender: *address.name(),
-            message: Rpc::Response {
-                requester,
-                response: Response::Mutation(result),
-                message_id,
-            },
-        })
-    }
-
-    /// Put MData.
-    fn handle_put_mdata_req(
-        &mut self,
-        requester: PublicId,
-        data: MData,
-        message_id: MessageId,
-    ) -> Option<Action> {
-        let result = if self.mutable_chunks.has(data.address()) {
-            Err(NdError::DataExists)
-        } else {
-            self.mutable_chunks
-                .put(&data)
-                .map_err(|error| error.to_string().into())
-        };
-        Some(Action::RespondToClientHandlers {
-            sender: *data.name(),
-            message: Rpc::Response {
-                requester,
-                response: Response::Mutation(result),
-                message_id,
-            },
-        })
-    }
-
-    fn handle_delete_mdata_req(
-        &mut self,
-        requester: PublicId,
-        address: MDataAddress,
-        message_id: MessageId,
-    ) -> Option<Action> {
-        let requester_pk = *utils::own_key(&requester)?;
-
-        let result = self
-            .mutable_chunks
-            .get(&address)
-            .map_err(|e| match e {
-                ChunkStoreError::NoSuchChunk => NdError::NoSuchData,
-                error => error.to_string().into(),
-            })
-            .and_then(move |mdata| {
-                mdata.check_is_owner(requester_pk)?;
-
-                self.mutable_chunks
-                    .delete(&address)
-                    .map_err(|error| error.to_string().into())
-            });
-
-        Some(Action::RespondToClientHandlers {
-            sender: *address.name(),
-            message: Rpc::Response {
-                requester,
-                response: Response::Mutation(result),
-                message_id,
-            },
-        })
-    }
-
-    /// Set MData user permissions.
-    fn handle_set_mdata_user_permissions_req(
-        &mut self,
-        requester: PublicId,
-        address: MDataAddress,
-        user: PublicKey,
-        permissions: &MDataPermissionSet,
-        version: u64,
-        message_id: MessageId,
-    ) -> Option<Action> {
-        let requester_pk = *utils::own_key(&requester)?;
-
-        self.mutate_mdata_chunk(&address, requester, message_id, move |mut data| {
-            data.check_permissions(MDataAction::ManagePermissions, requester_pk)?;
-            data.set_user_permissions(user, permissions.clone(), version)?;
-            Ok(data)
-        })
-    }
-
-    /// Delete MData user permissions.
-    fn handle_del_mdata_user_permissions_req(
-        &mut self,
-        requester: PublicId,
-        address: MDataAddress,
-        user: PublicKey,
-        version: u64,
-        message_id: MessageId,
-    ) -> Option<Action> {
-        let requester_pk = *utils::own_key(&requester)?;
-
-        self.mutate_mdata_chunk(&address, requester, message_id, move |mut data| {
-            data.check_permissions(MDataAction::ManagePermissions, requester_pk)?;
-            data.del_user_permissions(user, version)?;
-            Ok(data)
-        })
-    }
-
-    /// Mutate Sequenced MData.
-    fn handle_mutate_seq_mdata_entries_req(
-        &mut self,
-        requester: PublicId,
-        address: MDataAddress,
-        actions: MDataSeqEntryActions,
-        message_id: MessageId,
-    ) -> Option<Action> {
-        let requester_pk = *utils::own_key(&requester)?;
-
-        self.mutate_mdata_chunk(&address, requester, message_id, move |mut data| {
-            match data {
-                MData::Seq(ref mut mdata) => mdata.mutate_entries(actions, requester_pk)?,
-                MData::Unseq(..) => {
-                    error!("Logic error - unexpected chunk stored at {:?}", address);
-                    return Err(NdError::NetworkOther(
-                        "Logic error - unexpected chunk".to_string(),
-                    ));
-                }
-            }
-            Ok(data)
-        })
-    }
-
-    /// Mutate Unsequenced MData.
-    fn handle_mutate_unseq_mdata_entries_req(
-        &mut self,
-        requester: PublicId,
-        address: MDataAddress,
-        actions: MDataUnseqEntryActions,
-        message_id: MessageId,
-    ) -> Option<Action> {
-        let requester_pk = *utils::own_key(&requester)?;
-
-        self.mutate_mdata_chunk(&address, requester, message_id, move |mut data| {
-            match data {
-                MData::Unseq(ref mut mdata) => mdata.mutate_entries(actions, requester_pk)?,
-                MData::Seq(..) => {
-                    error!("Logic error - unexpected chunk stored at {:?}", address);
-                    return Err(NdError::NetworkOther(
-                        "Logic error - unexpected chunk".to_string(),
-                    ));
-                }
-            }
-            Ok(data)
-        })
-    }
-
-    /// Get entire MData.
-    fn handle_get_mdata_req(
-        &mut self,
-        requester: PublicId,
-        address: MDataAddress,
-        message_id: MessageId,
-    ) -> Option<Action> {
-        let result = self.get_mdata_chunk(&address, &requester, MDataAction::Read)?;
-
-        Some(Action::RespondToClientHandlers {
-            sender: *address.name(),
-            message: Rpc::Response {
-                requester,
-                response: Response::GetMData(result),
-                message_id,
-            },
-        })
-    }
-
-    /// Get MData shell.
-    fn handle_get_mdata_shell_req(
-        &mut self,
-        requester: PublicId,
-        address: MDataAddress,
-        message_id: MessageId,
-    ) -> Option<Action> {
-        let result = self
-            .get_mdata_chunk(&address, &requester, MDataAction::Read)?
-            .map(|data| data.shell());
-
-        Some(Action::RespondToClientHandlers {
-            sender: *address.name(),
-            message: Rpc::Response {
-                requester,
-                response: Response::GetMDataShell(result),
-                message_id,
-            },
-        })
-    }
-
-    /// Get MData version.
-    fn handle_get_mdata_version_req(
-        &mut self,
-        requester: PublicId,
-        address: MDataAddress,
-        message_id: MessageId,
-    ) -> Option<Action> {
-        let result = self
-            .get_mdata_chunk(&address, &requester, MDataAction::Read)?
-            .map(|data| data.version());
-
-        Some(Action::RespondToClientHandlers {
-            sender: *address.name(),
-            message: Rpc::Response {
-                requester,
-                response: Response::GetMDataVersion(result),
-                message_id,
-            },
-        })
-    }
-
-    /// Get MData value.
-    fn handle_get_mdata_value_req(
-        &mut self,
-        requester: PublicId,
-        address: MDataAddress,
-        key: &[u8],
-        message_id: MessageId,
-    ) -> Option<Action> {
-        let res = self.get_mdata_chunk(&address, &requester, MDataAction::Read)?;
-
-        let response = if address.is_seq() {
-            Response::GetSeqMDataValue(res.and_then(|data| match data {
-                MData::Seq(md) => md.get(key).cloned().ok_or_else(|| NdError::NoSuchEntry),
-                MData::Unseq(..) => {
-                    error!("Logic error - unexpected chunk stored at {:?}", address);
-                    Err(NdError::NetworkOther(
-                        "Logic error - unexpected chunk".to_string(),
-                    ))
-                }
-            }))
-        } else {
-            Response::GetUnseqMDataValue(res.and_then(|data| match data {
-                MData::Unseq(md) => md.get(key).cloned().ok_or_else(|| NdError::NoSuchEntry),
-                MData::Seq(..) => {
-                    error!("Logic error - unexpected chunk stored at {:?}", address);
-                    Err(NdError::NetworkOther(
-                        "Logic error - unexpected chunk".to_string(),
-                    ))
-                }
-            }))
-        };
-
-        Some(Action::RespondToClientHandlers {
-            sender: *address.name(),
-            message: Rpc::Response {
-                requester,
-                response,
-                message_id,
-            },
-        })
-    }
-
-    /// Get MData keys.
-    fn handle_list_mdata_keys_req(
-        &mut self,
-        requester: PublicId,
-        address: MDataAddress,
-        message_id: MessageId,
-    ) -> Option<Action> {
-        let result = self
-            .get_mdata_chunk(&address, &requester, MDataAction::Read)?
-            .map(|data| data.keys());
-
-        Some(Action::RespondToClientHandlers {
-            sender: *address.name(),
-            message: Rpc::Response {
-                requester,
-                response: Response::ListMDataKeys(result),
-                message_id,
-            },
-        })
-    }
-
-    /// Get MData values.
-    fn handle_list_mdata_values_req(
-        &mut self,
-        requester: PublicId,
-        address: MDataAddress,
-        message_id: MessageId,
-    ) -> Option<Action> {
-        let res = self.get_mdata_chunk(&address, &requester, MDataAction::Read)?;
-
-        let response = if address.is_seq() {
-            Response::ListSeqMDataValues(res.and_then(|data| match data {
-                MData::Seq(md) => Ok(md.values()),
-                MData::Unseq(..) => {
-                    error!("Logic error - unexpected chunk stored at {:?}", address);
-                    Err(NdError::NetworkOther(
-                        "Logic error - unexpected chunk".to_string(),
-                    ))
-                }
-            }))
-        } else {
-            Response::ListUnseqMDataValues(res.and_then(|data| match data {
-                MData::Unseq(md) => Ok(md.values()),
-                MData::Seq(..) => {
-                    error!("Logic error - unexpected chunk stored at {:?}", address);
-                    Err(NdError::NetworkOther(
-                        "Logic error - unexpected chunk".to_string(),
-                    ))
-                }
-            }))
-        };
-
-        Some(Action::RespondToClientHandlers {
-            sender: *address.name(),
-            message: Rpc::Response {
-                requester,
-                response,
-                message_id,
-            },
-        })
-    }
-
-    /// Get MData entries.
-    fn handle_list_mdata_entries_req(
-        &mut self,
-        requester: PublicId,
-        address: MDataAddress,
-        message_id: MessageId,
-    ) -> Option<Action> {
-        let res = self.get_mdata_chunk(&address, &requester, MDataAction::Read)?;
-
-        let response = if address.is_seq() {
-            Response::ListSeqMDataEntries(res.and_then(|data| match data {
-                MData::Seq(md) => Ok(md.entries().clone()),
-                MData::Unseq(..) => {
-                    error!("Logic error - unexpected chunk stored at {:?}", address);
-                    Err(NdError::NetworkOther(
-                        "Logic error - unexpected chunk".to_string(),
-                    ))
-                }
-            }))
-        } else {
-            Response::ListUnseqMDataEntries(res.and_then(|data| match data {
-                MData::Unseq(md) => Ok(md.entries().clone()),
-                MData::Seq(..) => {
-                    error!("Logic error - unexpected chunk stored at {:?}", address);
-                    Err(NdError::NetworkOther(
-                        "Logic error - unexpected chunk".to_string(),
-                    ))
-                }
-            }))
-        };
-
-        Some(Action::RespondToClientHandlers {
-            sender: *address.name(),
-            message: Rpc::Response {
-                requester,
-                response,
-                message_id,
-            },
-        })
-    }
-
-    /// Get MData permissions.
-    fn handle_list_mdata_permissions_req(
-        &mut self,
-        requester: PublicId,
-        address: MDataAddress,
-        message_id: MessageId,
-    ) -> Option<Action> {
-        let result = self
-            .get_mdata_chunk(&address, &requester, MDataAction::Read)?
-            .map(|data| data.permissions());
-
-        Some(Action::RespondToClientHandlers {
-            sender: *address.name(),
-            message: Rpc::Response {
-                requester,
-                response: Response::ListMDataPermissions(result),
-                message_id,
-            },
-        })
-    }
-
-    /// Get MData user permissions.
-    fn handle_list_mdata_user_permissions_req(
-        &mut self,
-        requester: PublicId,
-        address: MDataAddress,
-        user: PublicKey,
-        message_id: MessageId,
-    ) -> Option<Action> {
-        let result = self
-            .get_mdata_chunk(&address, &requester, MDataAction::Read)?
-            .and_then(|data| data.user_permissions(user).map(MDataPermissionSet::clone));
-
-        Some(Action::RespondToClientHandlers {
-            sender: *address.name(),
-            message: Rpc::Response {
-                requester,
-                response: Response::ListMDataUserPermissions(result),
-                message_id,
-            },
-        })
     }
 
     fn handle_put_idata_req(
